Make PCI claim NUMA aware during live migration
NUMA aware live migration and SRIOV live migration was implemented as
two separate feature. As a consequence the case when both SRIOV and NUMA
is present in the instance was missed. When the PCI device is claimed on
the destination host the NUMA topology of the instance needs to be
passed to the claim call.

Change-Id: If469762b22d687151198468f0291821cebdf26b2
Closes-Bug: #1893221