Allow tap interface with multiqueue
When vif_type="tap" (such as when using calico),
attempting to create an instance using an image that has
the property hw_vif_multiqueue_enabled=True fails, because
the interface is always being created without multiqueue
flags.

This change checks if the property is defined and passes
the multiqueue parameter to create the tap interface
accordingly.

In case the multiqueue parameter is passed but the
vif_model is not virtio (or unspecified), the old
behavior is maintained.

Conflicts:
    nova/virt/libvirt/vif.py

NOTE: The conflict is due to not having patch
Iefa6009874395cafe3121e44a48d882c38230d16 in stable/ussuri.

Change-Id: I0307c43dcd0cace1620d2ac75925651d4ee2e96c
Closes-bug: #1893263
(cherry picked from commit 84cfc8e9ab1396ec17abcfc9646c7d40f1d966ae)