Handle oslo.messaging ping endpoint
We have a test that attempts to emulate a messaging endpoint and fails
if anything other than the endpoint's target property is inspected. With
the addition of the ping endpoint in
I51cf67e060f240e6eb82260e70a057fe599f9063, this now blows up with the
latest oslo.messagin release.

To handle this situation, our fake class just needs to also check for
the oslo_rpc_server_ping attribute.

Change-Id: I970c953f9c21df99fac5f572f9e9dca5c615f06c
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>