zuul: Replace nova-live-migration with zuulv3 jobs
This change removes the original nova-live-migration job and replaces it
directly with the new Focal based zuulv3 native job.

The nova-dsvm-multinode-base base job is no longer used and so also
removed as part of this change.

Note that this new nova-live-migration job does not yet contain any
ceph coverage like the original, this is still pending and will be
completed early in the W cycle.

This change is being merged ahead of this to resolve bug #1901739, a
known QEMU -drive issue caused by the previous jobs use of libvirt 5.4.0
as provided by Ubuntu Bionic. The fix here being the migration to Ubuntu
Focal based jobs and libvirt 6.0.0 that now defaults to using QEMU
-blockdev.

Closes-Bug: #1901739
Change-Id: Ib342e2d3c395830b4667a60de7e492d3b9de2f0a