neutron: Remove references to dead 'qos-queue' extension
We had assumed that this was used by the out-of-tree XenAPI driver, but
it turns out to have been used by the VMWare NSX QoS extension [1]. This
extension itself is now EOL and the code was removed from the vmware-nsx
package over two years ago [2][3]. There are no remaining references to
this former extension:

  $ cd vmware-nsx
  $ ag qos-queue
  $ git log -S qos-queue --oneline | head -1
  26135f34a retire the NSX MH plugin

We can safely remove this. A release note is included even though this
is a vendor-specific feature that should really never have been merged.

[1] https://github.com/openstack/neutron/blob/fe61e29f/doc/source/admin/archives/adv-features.rst#vmware-nsx-qos-extension
[2] https://opendev.org/x/vmware-nsx/src/commit/b5f59ece9130016f8949ac1b6c7a242d0048795f/vmware_nsx/extensions/qos_queue.py
[3] https://opendev.org/x/vmware-nsx/commit/26135f34acca66ce994e0dd4280e2e968ed10ce3

Change-Id: I90171131d671d1b9ef2f93cc13174ebb00e96033
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>