functional: Add tests for mixed CPU policy
I'm not sure why these weren't included in the series, but they help
prove out what this is doing.

This highlights a small "bug" in the code, whereby the topology object
associated with the instance's NUMA cell doesn't have the correct number
of CPUs. This has no adverse effects since that attribute isn't actually
used except to indicate a minimum thread count necessary for the cell,
but it is wrong as we fix it all the same.

Some tests are renamed now that we're testing more than one "legacy"
policy.

Part of blueprint use-pcpu-and-vcpu-in-one-instance

Change-Id: I30a040d549f48d53cab2c59c00bb269f821ace88
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>