functional: Add test for #1898272
If you have two host NUMA nodes and you mark the entirety of one node
for shared vCPUs (VCPU) and the other for dedicated vCPUs (PCPU), then
it should not be possible to boot an instance with a 'mixed' CPU policy
and a single NUMA node, since satisfying that request should necessitate
splitting across host nodes. This is not currently the case. Prove that
out.

Change-Id: I72afe21d117e932cc00c9f616618fb3722bf8ca4
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Related-Bug: #1898272