libvirt: Only add a USB controller if it's necessary
Libvirt will add a USB controller by default on x86-64 hosts. This is
legacy behavior due to QEMU's historical behaviour of not providing a
way to disable them. However, a USB controller is both unnecessary and
in some cases (real-time) actively harmful if there are no guest devices
that rely on the USB bus.

The libvirt documentation suggests that there are a number of devices
that can use a USB bus: input devices, various types of disk device, hub
devices, USB serial ports, sound devices, and host passthrough and
redirected USB devices. Of these, nova supports only the first two. Add
a check for these devices and if present, add a controller, else disable
the controller.

Change-Id: I5b16d9ddc9cad56b367d1ad94abba95d675840b0
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>