Move query param generation to RequestGroup o.vo
This patch moves the allocation candidate query generation code from the
scheduler utils to the RequestGroup o.vo as it clearly only depends on
that object. The original utils code was predated the patch that added
the RequestGroup o.vo

There is small change in logic. The mapping of the unnamed group, that
does not have a specific requester, to the empty query prefix is now
moved into the query param generation code.

Change-Id: I2fb5b6a8fdbb6c13554e0608b71a02843d12d484
blueprint: support-interface-attach-with-qos-ports