Factor out port resource deallocation
During detach, and later during a failed attach, the port resource
allocation needs to be deleted from placement. Currently this is done
together with the unbinding of the port. This patch creates a separate
function for resource deallocation. This will be used later separately
when an qos interface attach fails after the resource allocation but
before the port is bound in neutron.

Change-Id: If3e79cd71b6d0f6e535ff86b27483c137a78fee9
blueprint: support-interface-attach-with-qos-ports