Refactor update_pci_request_spec_with_allocated_interface_name
Make update_pci_request_spec_with_allocated_interface_name only depend
on a list of IntancePCIRequest o.vos instead of a whole Instance object.
This will come in handy for the qos interface attach case where we only
need to make the changes on the Instance o.vo after we are sure that
the both the resource allocation and the pci claim is succeeded for the
request.

Change-Id: I5a6c6d3eed61895b00f9e9c3fb3b5d09d6786e9c
blueprint: support-interface-attach-with-qos-ports